# [Bug Fix] oom: pd repeate to compute mm encoder